RM-1127: Tweaking this to provide better shared functionality.
