+++ lib/spark_api/models/virtual_tour.rb
@@ -2,7 +2,7 @@ module SparkApi
   module Models
     class VirtualTour < Base
       extend Subresource
-      include MediaPrivacy
+      include Media
       include Concerns::Savable,
               Concerns::Destroyable
 
@@ -15,6 +15,14 @@ module SparkApi
 
       def unbranded?
         attributes["Type"] == "unbranded"
+      end
+
+      def url
+        attributes['Uri']
+      end
+
+      def description
+        attributes['Name']
       end
     end
   end
